+++ packages/shadydom/CHANGELOG.md
@@ -9,6 +9,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+- Adds improvements to `noPatch: 'on-demand'` mode, including:
+  `ShadyDOM.validateNodePatch` which can reset a patch on a node if needed,
+  `ShadyDOM.onDemandGlobalPatches` which is a list of properties to always patch,
+  always patches styling related properties, and allows a limited form of
+  overriding native methods.
+  ([#484](https://github.com/webcomponents/polyfills/pull/484))
 - Add type annotations for JSCompiler to `eventPhase` property descriptor.
   ([#473](https://github.com/webcomponents/polyfills/pull/473))
 - Allow event listener options to be specified using a function in addition to